% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotStar.R
\name{plotStar}
\alias{plotStar}
\title{Plot a circle of radius r.}
\usage{
plotStar(r = 1)
}
\arguments{
\item{r}{The radius of the circle to
  be plotted}
}
\value{
None The function is called
  solely for its side effect
}
\description{
Normally, we would now describe what
the function is supposed to do
}
\examples{
plotStar(10)
}
\seealso{
The general
  \code{\link[graphics]{plot}} function
}

